use int
Former-commit-id: 4335e32f3ea9d6e3254750b516f42eb3fe913e0d

+++ test/functional/feature_asset_mint.py
@@ -61,7 +61,7 @@ class AssetMintTest(SyscoinTestFramework):
         assert_raises_rpc_error(-4, 'mint-exists', self.nodes[0].assetallocationmint, self.asset, newaddress, '100', height, bridgetransferid, spv_tx_value, spv_tx_parent_nodes, spv_tx_path, spv_receipt_value, spv_receipt_parent_nodes)
         
         # increase time by ~2.5 week and assetallocationmint should throw timeout error
-        numBlocks = 1513000 / (2*60*59)
+        numBlocks = int(1513000 / (2*60*59))
         for block in range(numBlocks):
             set_node_times(self.nodes, self.nodes[0].getblockheader(self.nodes[0].getbestblockhash())["time"] + (2*60*59))
             self.nodes[0].generate(1)
